% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/load_data.R
\name{load_data}
\alias{load_data}
\title{Load ABC Atlas data}
\usage{
load_data(download_base = "../../abc_download_root")
}
\arguments{
\item{download_base}{The base directory where the data is downloaded.}
}
\value{
A list containing cell metadata and gene data.
}
\description{
This function loads the ABC Atlas data from the specified download base directory.
}
