% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/load_data_HMBA.R
\name{load_data_HMBA}
\alias{load_data_HMBA}
\title{Load ABC Atlas HMBA Multiome data (HMBA-10xMultiome-BG-Aligned)}
\usage{
load_data_HMBA(download_base = "abc_download_root")
}
\arguments{
\item{download_base}{The base directory where the data is downloaded.}
}
\value{
A list containing extended cell metadata, donor, library, and gene data.
}
\description{
This function loads the HMBA ABC Atlas data and joins cell, donor, and library metadata.
}
