% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/load_data_PMDBS.R
\name{load_data_PMDBS}
\alias{load_data_PMDBS}
\title{Load ABC Atlas PMDBS data}
\usage{
load_data_PMDBS(download_base = "abc_download_root")
}
\arguments{
\item{download_base}{The base directory where the data is downloaded.}
}
\value{
A list containing cell metadata, gene data, and other metadata tables.
}
\description{
This function loads the ABC Atlas PMDBS data from the specified download base directory.
}
