% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/load_data_WHB.R
\name{load_data_WHB}
\alias{load_data_WHB}
\title{Load ABC Atlas WHB data}
\usage{
load_data_WHB(download_base = "abc_download_root")
}
\arguments{
\item{download_base}{The base directory where the data is downloaded.}
}
\value{
A list containing cell metadata and gene data.
}
\description{
This function loads the ABC Atlas WHB data from the specified download base directory.
}
