% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/load_data_WMB.R
\name{load_data_WMB}
\alias{load_data_WMB}
\title{Load ABC Atlas WMB data}
\usage{
load_data_WMB(download_base = "abc_download_root")
}
\arguments{
\item{download_base}{The base directory where the data is downloaded.}
}
\value{
A list containing cell metadata and gene data.
}
\description{
This function loads the ABC Atlas WMB data from the specified download base directory.
}
