% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/setup.R
\name{setup_environment}
\alias{setup_environment}
\title{Setup the Python environment}
\usage{
setup_environment()
}
\description{
This function sets up the Python environment required to run the data loading functions.
}
